server/requirements.txt
fastapi
uvicorn
torch
numpy
Pillow
realesrgan
basicsr
